+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * keep lines with location in summary mode - Fix #11
 * allow scrolling the report (arrow keys, page keys, home & end keys, mouse wheel) - Fix #6
 * log file renamed to 'bacon.log' to avoid collisions
+* initial execution is displayed raw before report computation
 
 <a name="v0.2.0"></a>
 ### v0.2.0 - 2020-10-01